pyproject.toml
[project]
name = "life-hacker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "openai-agents>=0.1.0",
]
